+++ __tests__/setup/pageObjects/endSpeechSynthesize.js
@@ -1,0 +1,5 @@
+import executePromiseScript from './executePromiseScript';
+
+export default async function endSpeechSynthesize(driver) {
+  return await executePromiseScript(driver, () => window.WebSpeechMock.mockEndSynthesize());
+}