+++ __tests__/setup/pageObjects/index.js
@@ -1,6 +1,14 @@
 import dispatchAction from './dispatchAction';
+import endSpeechSynthesize from './endSpeechSynthesize';
+import executePromiseScript from './executePromiseScript';
+import getMicrophoneButton from './getMicrophoneButton';
+import getSendBoxTextBox from './getSendBoxTextBox';
+import hasPendingSpeechSynthesisUtterance from './hasPendingSpeechSynthesisUtterance';
+import isRecognizingSpeech from './isRecognizingSpeech';
 import pingBot from './pingBot';
+import putSpeechRecognitionResult from './putSpeechRecognitionResult';
 import sendMessageViaSendBox from './sendMessageViaSendBox';
+import startSpeechSynthesize from './startSpeechSynthesize';
 
 function mapMap(map, mapper) {
   return Object.keys(map).reduce((final, key) => {
@@ -14,8 +22,16 @@ export default function pageObjects(driver) {
   return mapMap(
     {
       dispatchAction,
+      endSpeechSynthesize,
+      executePromiseScript,
+      getMicrophoneButton,
+      getSendBoxTextBox,
+      hasPendingSpeechSynthesisUtterance,
+      isRecognizingSpeech,
       pingBot,
-      sendMessageViaSendBox
+      putSpeechRecognitionResult,
+      sendMessageViaSendBox,
+      startSpeechSynthesize
     },
     fn => fn.bind(null, driver)
   );