+++ __tests__/setup/pageObjects/isRecognizingSpeech.js
@@ -1,0 +1,3 @@
+export default async function isRecognizingSpeech(driver) {
+  return await driver.executeScript(() => window.WebSpeechMock.isRecognizing());
+}