+++ __tests__/setup/pageObjects/sendMessageViaSendBox.js
@@ -1,10 +1,11 @@
-import { By, Key } from 'selenium-webdriver';
+import { Key } from 'selenium-webdriver';
 
 import { timeouts } from '../../constants.json';
 import allOutgoingActivitiesSent from '../conditions/allOutgoingActivitiesSent';
+import getSendBoxTextBox from './getSendBoxTextBox';
 
 export default async function sendMessageViaSendBox(driver, text, { waitForSend = true }) {
-  const input = await driver.findElement(By.css('[role="form"] > * > form > input[type="text"]'));
+  const input = await getSendBoxTextBox(driver);
 
   await input.sendKeys(text, Key.RETURN);